Update utils.js to detect iPadOS
Since iOS 13, the iPad Pro runs under iPadOS and does not identify itself as a iDevice anymore. There have been a lot of approaches to identify an iPadOS but they all are not failsafe. My current method taken from https://stackoverflow.com/a/59049628/3297914 has been pretty good and so far failsafe on iPadOS 13 and 14.

Things might change when Apple will release a touch display anytime in the future, but for know, this check adds solid detection for all iPad (Pros) running iPadOS 13.
Fixes issues on iPad.

The check `&& typeof navigator.standalone !== "undefined"` is only true on Safari MAC, since its the only browser that uses navigator.standalone, so this check also should be safe in the future, when touch displays enter the market. Since I use BSL in every project, I have a tight eye on regressions and will continue to update.